refactor(edgeless): using common createButtonPopper
