+++ packages/blocks/src/page-block/edgeless/components/utils.ts
@@ -2,6 +2,7 @@ import type { Bound } from '@blocksuite/phasor';
 import { deserializeXYWH, type SurfaceViewport } from '@blocksuite/phasor';
 import { getCommonBound } from '@blocksuite/phasor';
 import type { Disposable } from '@blocksuite/store';
+import { createPopper } from '@popperjs/core';
 
 import type { Selectable } from '../selection-manager.js';
 import { getSelectionBoxBound, getXYWH, isTopLevelBlock } from '../utils.js';
@@ -88,6 +89,81 @@ export function listenClickAway(
   return {
     dispose: () => {
       document.removeEventListener('click', callback);
+    },
+  };
+}
+
+const ATTR_SHOW = 'data-show';
+/**
+ * Using attribute 'data-show' to control popper visibility.
+ *
+ * ```css
+ * selector {
+ *   display: none;
+ * }
+ * selector[data-show] {
+ *   display: block;
+ * }
+ * ```
+ */
+export function createButtonPopper(
+  reference: HTMLElement,
+  popperElement: HTMLElement
+) {
+  const popper = createPopper(reference, popperElement, {
+    placement: 'top',
+    modifiers: [
+      {
+        name: 'offset',
+        options: {
+          offset: [0, 12],
+        },
+      },
+    ],
+  });
+
+  const show = () => {
+    popperElement.setAttribute(ATTR_SHOW, '');
+    popper.setOptions(options => ({
+      ...options,
+      modifiers: [
+        ...(options.modifiers ?? []),
+        { name: 'eventListeners', enabled: false },
+      ],
+    }));
+    popper.update();
+  };
+
+  const hide = () => {
+    popperElement.removeAttribute(ATTR_SHOW);
+
+    popper.setOptions(options => ({
+      ...options,
+      modifiers: [
+        ...(options.modifiers ?? []),
+        { name: 'eventListeners', enabled: false },
+      ],
+    }));
+  };
+
+  const toggle = () => {
+    if (popperElement.hasAttribute(ATTR_SHOW)) {
+      hide();
+    } else {
+      show();
+    }
+  };
+
+  const clickAway = listenClickAway(reference, () => hide());
+
+  return {
+    popper,
+    show,
+    hide,
+    toggle,
+    dispose: () => {
+      popper.destroy();
+      clickAway.dispose();
     },
   };
 }
