+++ packages/blocks/src/page-block/edgeless/toolbar/shape-tool/shape-tool-button.ts
@@ -3,74 +3,16 @@ import './shape-menu.js';
 
 import { ShapeIcon } from '@blocksuite/global/config';
 import { DisposableGroup } from '@blocksuite/store';
-import { createPopper } from '@popperjs/core';
 import { css, html, LitElement } from 'lit';
 import { customElement, property, query } from 'lit/decorators.js';
 
-import type { MouseMode } from '../../../../__internal__/index.js';
+import type {
+  MouseMode,
+  ShapeMouseMode,
+} from '../../../../__internal__/index.js';
+import { createButtonPopper } from '../../components/utils.js';
 import type { EdgelessPageBlockComponent } from '../../edgeless-page-block.js';
 import type { EdgelessShapeMenu } from './shape-menu.js';
-
-const ATTR_SHOW = 'data-show';
-
-function createShapeMenuPopper(
-  reference: HTMLElement,
-  popperElement: HTMLElement
-) {
-  const popper = createPopper(reference, popperElement, {
-    placement: 'top',
-    modifiers: [
-      {
-        name: 'offset',
-        options: {
-          offset: [0, 12],
-        },
-      },
-    ],
-  });
-
-  const show = () => {
-    popperElement.setAttribute(ATTR_SHOW, '');
-    popper.setOptions(options => ({
-      ...options,
-      modifiers: [
-        ...(options.modifiers ?? []),
-        { name: 'eventListeners', enabled: false },
-      ],
-    }));
-    popper.update();
-  };
-
-  const hide = () => {
-    popperElement.removeAttribute(ATTR_SHOW);
-
-    popper.setOptions(options => ({
-      ...options,
-      modifiers: [
-        ...(options.modifiers ?? []),
-        { name: 'eventListeners', enabled: false },
-      ],
-    }));
-  };
-
-  const toggle = () => {
-    if (popperElement.hasAttribute(ATTR_SHOW)) {
-      hide();
-    } else {
-      show();
-    }
-  };
-
-  return {
-    popper,
-    show,
-    hide,
-    toggle,
-    dispose: () => {
-      popper.destroy();
-    },
-  };
-}
 
 @customElement('edgeless-shape-tool-button')
 export class EdgelessShapeToolButton extends LitElement {
@@ -97,8 +39,7 @@ export class EdgelessShapeToolButton extends LitElement {
   @query('edgeless-shape-menu')
   private _shapeMenu!: EdgelessShapeMenu;
 
-  private _shapeMenuPopper: ReturnType<typeof createShapeMenuPopper> | null =
-    null;
+  private _shapeMenuPopper: ReturnType<typeof createButtonPopper> | null = null;
 
   private _disposables: DisposableGroup = new DisposableGroup();
 
@@ -106,14 +47,18 @@ export class EdgelessShapeToolButton extends LitElement {
     this._shapeMenuPopper?.toggle();
   }
 
+  private _setMouseMode(mode: ShapeMouseMode) {
+    this.edgeless.slots.mouseModeUpdated.emit(mode);
+  }
+
   firstUpdated(changedProperties: Map<string, unknown>) {
     const _disposables = this._disposables;
 
-    this._shapeMenuPopper = createShapeMenuPopper(this, this._shapeMenu);
+    this._shapeMenuPopper = createButtonPopper(this, this._shapeMenu);
     _disposables.add(this._shapeMenuPopper);
     _disposables.add(
       this._shapeMenu.slots.select.on(shape => {
-        this.edgeless.slots.mouseModeUpdated.emit({
+        this._setMouseMode({
           type: 'shape',
           shape,
           color: '#000000',
@@ -144,7 +89,7 @@ export class EdgelessShapeToolButton extends LitElement {
         .tooltip=${'Shape'}
         .active=${type === 'shape'}
         .testId=${'shape'}
-        @tool.click=${() => this._toggleShapeMenu()}
+        @tool.click=${async () => this._toggleShapeMenu()}
       >
         ${ShapeIcon}
       </edgeless-tool-icon-button>
